fix: dedupe GettingStartedAction command map across Main and Progress views
MainViewMessageHandler and ProgressViewMessageHandler each hardcoded
their own copy of the runSetup/createSampleProject/cloneOverleaf/
downloadArxiv/openWalkthrough -> command ID map for the same getting
started actions. Nothing kept the two literals in sync, so a rename
of one command ID would silently break only one of the two surfaces.
Move the map to a single shared constant (GETTING_STARTED_COMMANDS in
src/shared/schemas/mainView/state.ts) and import it from both handlers.
