+++ packages/extension/src/progressView/ProgressViewMessageHandler.ts
@@ -33,7 +33,11 @@ import type { PromptHost } from '@hosts/uiHosts';
 import { isApiProvider } from '@model/apiProviders';
 import { invalidateModelOptionsCache } from '@model/computeModelOptions';
 import { COMMON_COMMANDS, PROGRESS_VIEW_COMMANDS } from '@shared/ipc';
-import type { GettingStartedAction, StreamTabId } from '@shared/schemas';
+import {
+  GETTING_STARTED_COMMANDS,
+  type GettingStartedAction,
+  type StreamTabId,
+} from '@shared/schemas';
 import { isGoalInFlight } from '@shared/schemas/goal';
 import {
   dispatchProgressViewInbound,
@@ -297,15 +301,11 @@ export class ProgressViewMessageHandler extends BaseViewMessageHandler<
   private async runGettingStartedAction(
     action: GettingStartedAction,
   ): Promise<void> {
-    const commands = {
-      runSetup: 'texra.runSetupAssistant',
-      createSampleProject: 'texra.createSampleProject',
-      cloneOverleaf: 'texra.cloneOverleafProject',
-      downloadArxiv: 'texra.downloadArXivSource',
-      openWalkthrough: 'texra.openGettingStarted',
-    } satisfies Record<GettingStartedAction, string>;
-
-    await safeExecuteCommand(commands[action], [], this.viewName);
+    await safeExecuteCommand(
+      GETTING_STARTED_COMMANDS[action],
+      [],
+      this.viewName,
+    );
     if (action === 'runSetup') {
       await this.provider.refreshOnboardingFunnel();
     }
